docs: add fastly namespaced modules to the types
