+++ test-d/fastly:config-store.test-d.ts
@@ -1,0 +1,12 @@
+/// <reference path="../types/fastly:config-store.d.ts" />
+import { ConfigStore } from "fastly:config-store";
+import {expectError, expectType} from 'tsd';
+
+// ConfigStore
+{
+    expectError(new ConfigStore())
+    expectError(ConfigStore('example'))
+    expectError(ConfigStore())
+    expectType<ConfigStore>(new ConfigStore('example'))
+    expectType<(key:string) => string>(new ConfigStore('example').get)
+}
