+++ types/index.d.ts
@@ -1,0 +1,7 @@
+/// <reference path="fastly:backend.d.ts" />
+/// <reference path="fastly:config-store.d.ts" />
+/// <reference path="fastly:dictionary.d.ts" />
+/// <reference path="fastly:env.d.ts" />
+/// <reference path="fastly:geolocation.d.ts" />
+/// <reference path="fastly:object-store.d.ts" />
+/// <reference path="globals.d.ts" />
